+++ src/pages/MainLayout.tsx
@@ -1,16 +1,16 @@
-import React from 'react';
-import { Outlet } from 'react-router-dom';
-import NavBar from '../components/NavBar';
+import React from "react";
+import { Outlet } from "react-router-dom";
+import NavBar from "../components/NavBar";
 
 const MainLayout = () => {
   return (
     <>
       <NavBar />
-      <div>
+      <div className="p-3">
         <Outlet />
       </div>
     </>
-  )
+  );
 };
 
 export default MainLayout;